Move healthcheck to its own component
