+++ src/components/Home/Home.js
@@ -1,0 +1,7 @@
+import React from 'react';
+
+function Home() {
+  return <h2>You're home :)</h2>
+}
+
+export default Home;
